feat: remove bfx ui integration
this removes the algo-server module from the source code which was
uploading ui definitions / settings for the user to make them
available in the bitfinex main ui. it created a lot of additional
complexity and is running an additional, seperate websocket server
next to the 4 we are already using internally, and its a feature
not widely used, so we remove it to be able to focus on other
features.