remove exchange pool
this removes an additional websocket server that was responsible
for subscribing one websocket client to an exchange. probably the
initial intention was that when in a hosting scenario with many
different users, one separate server could manage data
subscriptions for different customers. usually this task would
have been solved by a message queue and not a custom websocket
server.

the change  simplifies the current code base, shrinks CPU and memory
load and therefore saves battery time for the HF users. it also
improves performance slightly as we have one middleware layer less.

this also prepares the code to support reconnect / reconnections
after a network disconnect.