refactor: move websocket emitter into the algo host
when we currently want to debug the algo host, we have to follow
code in many different projects. as we just provide one way to
connect algo orders, the plugin interface is doing more harm than
it is useful.

additionally it provides an additional layer of abstraction,
making it hard to use default api calls and events and slows down
the timely reaction of an algo order.

the main idea is to remove this additional layer with future
developements. having the event bus it provides co-located to the
actual algo orders makes refactoring and development easier.