add paper trading mode
this adds different modes for trading on a "main" account or a
"paper" account. old account data in the settings is migrated on
connect.

**login**: as usual login happens through the `auth.submit`-message.
the current mode, either `paper` or `main` must be defined now.
example to log into a paper trading subaccount:
`['auth.submit', 'SECRET', 'main']`

initially the server sends the configured modes those can be used
for the login screen in the UI, e.g. to show informations for a
conditional login. example:
`["info.modes_configured",["main","paper"]]` means that both a
paper and a main account is set up.

if a mode is not configured with api keys and the UI tries to login,
the api sends back informations that the api key is not
added:

```
['info.auth_api_key', false, $mode]
```

the above message means that no api key was found for a mode.

after login, the server sends the list of available markets for the
desired mode, as usual via the `info.markets` message.

**auth init**: when adding a new api key, the mode also has to get
specified, so the server can send the corresponding list of markets
available.

example:

on the start screen "paper" was selected. now api credentials are
added. for the example a hardcoded "paper" variable is used:

```js
dispatch(WSActions.send([
  'api_credentials.save',
  authToken,
  exID,
  apiKey,
  apiSecret,
  'paper',
]))
```