startup: recover when network resumes
in case the network is down or the initial request fails the
server does not start at all. it errors and print a stacktrace.

this adds a periodic retry in case the initial request to the bfx
api fails, e.g. due to a disconnected wifi.

to test, disconnect your internet connection. start the hf server.
connect internet again. the server should start within 3 seconds.