algos: pause and cleanup
when a user switches the app mode, the current running algo orders
should get pause. for that, the `onPause` command is added that
cleans up old state and stops the old algo server, without
removing the orders from the db.

additionally the `identify` message is called before `on_open` opens
the exchange connection for the algo host. at that point we don't
know the active instances yet. the `data.aos` event is moved to the
`on_open` handler.

the `on_open` handler for the conection will also clean up old,
orphaned algo order hosts. with a refactor that will remove the
algo server and integrates algos into the hf-api server directly,
thos state updates become easier to detect and manage.