support submit of multiple api keys
the ui displays both paper and main api keys in the settings now
and sends both to the server. this change supports a new parameter
that identifies the api key that was submitted.

additionally a reconnect to the api is just done when the app is
running in the mode the api key was submitted for.

support for https://github.com/bitfinexcom/bfx-hf-ui/pull/353